requirements.txt
annotated-types==0.7.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fhir.resources==7.1.0
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
